feat: add success page when register
