+++ src/shared/buttons/ButtonTFSecondary/ButtonTFSecondary.tsx
@@ -1,15 +1,17 @@
-
 type Props = {
-    onClick: () => void
-    label: string
+  onClick?: () => void
+  label: string
 }
 
 const ButtonTFSecondary = ({ onClick, label }: Props) => {
-    return (
-        <button onClick={onClick} className="font-sans text-p text-button-primary leading-8 cursor-pointer bg-iframe_bg_color py-13 px-30 rounded-100 hover:shadow-hover-btn focus:shadow-hover-btn active:shadow-hover-btn transition-btn-transition">
-            {label}
-        </button>
-    )
-};
+  return (
+    <button
+      onClick={onClick}
+      className="text-normal md:text-p text-button-primary leading-8 cursor-pointer py-13 px-30 rounded-100 hover:bg-hover-btn-secondary hover:shadow-hover-btn transition-all"
+    >
+      {label}
+    </button>
+  )
+}
 
-export default ButtonTFSecondary;
+export default ButtonTFSecondary
