+++ tailwind.config.js
@@ -9,6 +9,7 @@ module.exports = {
       'button-primary': '#FFB859',
       'hover-btn-primary': '#FF9C18',
       'button-secondary': '#503047',
+      'hover-btn-secondary': '#6750A414',
       'button-disabled': '#CAC4D0',
       error: '#CE2020',
       'text-color': '#292929',
@@ -45,9 +46,7 @@ module.exports = {
       spacing: {
         0: '0px',
         4: '4px',
-        5: '5px',
         8: '8px',
-        10: '10px',
         12: '12px',
         13: '13px',
         16: '16px',
@@ -62,6 +61,7 @@ module.exports = {
         60: '60px',
         70: '70px',
         100: '100px',
+        185: '185px',
         264: '264px',
         350: '350px',
         552: '552px',
